# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CreateShareLinkResponse": {
				"description": "Response when creating a share link",
				"properties": {
					"code": {
						"type": "string"
					},
					"shareUrl": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"shareUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"SharedWorkoutResponse": {
				"description": "Shared workout response returned by the API",
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"workoutName": {
						"type": "string"
					},
					"sessionInput": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"createdBy": {
						"properties": {
							"picture": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"name"
						],
						"type": "object"
					},
					"importCount": {
						"type": "number",
						"format": "double"
					},
					"expiresAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"code",
					"workoutName",
					"sessionInput",
					"generatedSession",
					"createdBy",
					"importCount",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImportSharedWorkoutResponse": {
				"description": "Response after importing a shared workout",
				"properties": {
					"workoutId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"workoutId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImportSharedWorkoutBody": {
				"description": "Body for importing a shared workout",
				"properties": {
					"name": {
						"type": "string",
						"description": "Optional custom name for the imported workout"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"WorkoutResponse": {
				"description": "Saved workout returned by the API",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sessionInput": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"sessionInput",
					"generatedSession",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateWorkoutBody": {
				"description": "Body for updating a saved workout",
				"properties": {
					"name": {
						"type": "string",
						"description": "New display name"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_",
						"description": "Updated generated session (when exercises are edited)"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SetData": {
				"description": "Set data for a single set within an exercise",
				"properties": {
					"reps": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"completed": {
						"type": "boolean"
					}
				},
				"required": [
					"reps",
					"completed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseLogData": {
				"description": "Exercise data logged during a session",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sets": {
						"items": {
							"$ref": "#/components/schemas/SetData"
						},
						"type": "array"
					},
					"isSetBased": {
						"type": "boolean"
					},
					"durationSeconds": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"sets",
					"isSetBased"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SessionLogResponse": {
				"description": "A completed training session log returned by the API",
				"properties": {
					"id": {
						"type": "string"
					},
					"workoutId": {
						"type": "string",
						"nullable": true
					},
					"sessionTitle": {
						"type": "string"
					},
					"sessionSubtitle": {
						"type": "string",
						"nullable": true
					},
					"startedAt": {
						"type": "string"
					},
					"durationSeconds": {
						"type": "number",
						"format": "double"
					},
					"exerciseCount": {
						"type": "number",
						"format": "double"
					},
					"notes": {
						"type": "string"
					},
					"exercises": {
						"items": {
							"$ref": "#/components/schemas/ExerciseLogData"
						},
						"type": "array",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"workoutId",
					"sessionTitle",
					"sessionSubtitle",
					"startedAt",
					"durationSeconds",
					"exerciseCount",
					"notes",
					"exercises",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSessionLogBody": {
				"description": "Body to create a new session log",
				"properties": {
					"workoutId": {
						"type": "string",
						"description": "Optional reference to a saved workout"
					},
					"sessionTitle": {
						"type": "string",
						"description": "Display title of the session"
					},
					"sessionSubtitle": {
						"type": "string",
						"description": "Subtitle (e.g. \"intermediate · 60 min · normal energy\")"
					},
					"startedAt": {
						"type": "string",
						"description": "ISO timestamp when the session started"
					},
					"durationSeconds": {
						"type": "number",
						"format": "double",
						"description": "Actual time spent in seconds"
					},
					"exerciseCount": {
						"type": "number",
						"format": "double",
						"description": "Total number of exercises performed"
					},
					"notes": {
						"type": "string",
						"description": "Free-form notes"
					},
					"exercises": {
						"items": {
							"$ref": "#/components/schemas/ExerciseLogData"
						},
						"type": "array",
						"description": "Detailed exercise data with sets/reps"
					}
				},
				"required": [
					"sessionTitle",
					"startedAt",
					"durationSeconds",
					"exerciseCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.ExerciseLogData_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/ExerciseLogData"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"ExerciseItem": {
				"description": "Exercise subset returned inside a generated session",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"focus": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"intensity": {
						"type": "number",
						"format": "double"
					},
					"defaultSets": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Default number of sets (null = time-based exercise)"
					},
					"defaultReps": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Default number of reps per set (null = time-based exercise)"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"category",
					"focus",
					"intensity",
					"defaultSets",
					"defaultReps"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SessionBlock": {
				"properties": {
					"phase": {
						"type": "string",
						"enum": [
							"warmup",
							"main",
							"addon",
							"cooldown"
						]
					},
					"phaseLabel": {
						"type": "string"
					},
					"exercises": {
						"items": {
							"properties": {
								"duration": {
									"type": "number",
									"format": "double"
								},
								"exercise": {
									"$ref": "#/components/schemas/ExerciseItem"
								}
							},
							"required": [
								"duration",
								"exercise"
							],
							"type": "object"
						},
						"type": "array"
					},
					"totalDuration": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"phase",
					"phaseLabel",
					"exercises",
					"totalDuration"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeneratedSession": {
				"properties": {
					"title": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"totalDuration": {
						"type": "number",
						"format": "double"
					},
					"blocks": {
						"items": {
							"$ref": "#/components/schemas/SessionBlock"
						},
						"type": "array"
					},
					"tips": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"subtitle",
					"totalDuration",
					"blocks",
					"tips"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Level": {
				"type": "string",
				"enum": [
					"beginner",
					"intermediate",
					"advanced"
				]
			},
			"Goal": {
				"type": "string",
				"enum": [
					"technique",
					"projecting",
					"power",
					"endurance",
					"volume",
					"recovery",
					"dynos",
					"handstand",
					"gym-push",
					"gym-pull",
					"gym-legs"
				]
			},
			"GymType": {
				"type": "string",
				"enum": [
					"slab",
					"comp",
					"spray",
					"moonboard",
					"mixed"
				]
			},
			"Fatigue": {
				"type": "string",
				"enum": [
					"fresh",
					"normal",
					"tired"
				]
			},
			"Equipment": {
				"type": "string",
				"enum": [
					"hangboard",
					"campus-board",
					"resistance-bands",
					"pull-up-bar",
					"none"
				]
			},
			"SessionInput": {
				"description": "Body sent to POST /api/sessions/generate",
				"properties": {
					"level": {
						"$ref": "#/components/schemas/Level"
					},
					"goal": {
						"$ref": "#/components/schemas/Goal"
					},
					"sessionLength": {
						"type": "number",
						"format": "double"
					},
					"gymType": {
						"$ref": "#/components/schemas/GymType"
					},
					"fatigue": {
						"$ref": "#/components/schemas/Fatigue"
					},
					"injuries": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"equipment": {
						"items": {
							"$ref": "#/components/schemas/Equipment"
						},
						"type": "array"
					}
				},
				"required": [
					"level",
					"goal",
					"sessionLength",
					"gymType",
					"fatigue",
					"injuries",
					"equipment"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "GRAVITACIO API",
		"version": "1.0.0",
		"description": "Bouldering workout planner API",
		"contact": {}
	},
	"paths": {
...
		"/api/shared/workouts/{workoutId}/share": {
			"post": {
				"operationId": "CreateShareLink",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateShareLinkResponse"
								}
							}
						}
					}
				},
				"description": "Create a share link for a workout. Requires authentication.\nThe workout must belong to the authenticated user.",
				"tags": [
					"Shared Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "workoutId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/shared/{code}": {
			"get": {
				"operationId": "GetSharedWorkout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SharedWorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Get details of a shared workout by code. No authentication required.",
				"tags": [
					"Shared Workouts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/shared/{code}/import": {
			"post": {
				"operationId": "ImportSharedWorkout",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImportSharedWorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Import a shared workout to the authenticated user's library.\nCreates a copy of the workout - doesn't link to the original.",
				"tags": [
					"Shared Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ImportSharedWorkoutBody"
							}
						}
					}
				}
			}
		},
		"/api/shared/my/links": {
			"get": {
				"operationId": "ListMyShareLinks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SharedWorkoutResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all share links created by the authenticated user.",
				"tags": [
					"Shared Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/workouts/{id}": {
			"get": {
				"operationId": "GetWorkout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Get a single saved workout by ID (must belong to the authenticated user).",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateWorkout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Update a saved workout (name and/or exercises).",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkoutBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWorkout",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Delete a saved workout.",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/session-logs": {
			"get": {
				"operationId": "ListSessionLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SessionLogResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List session logs for the authenticated user, newest first.\nOptionally filter by a start date (ISO string).",
				"tags": [
					"SessionLogs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "since",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSessionLog",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SessionLogResponse"
								}
							}
						}
					}
				},
				"description": "Record a completed training session.",
				"tags": [
					"SessionLogs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSessionLogBody"
							}
						}
					}
				}
			}
		},
		"/api/session-logs/{id}": {
			"delete": {
				"operationId": "DeleteSessionLog",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"description": "Delete a session log by ID.",
				"tags": [
					"SessionLogs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/session-logs/previous-exercises": {
			"post": {
				"operationId": "GetPreviousExerciseData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.ExerciseLogData_"
								}
							}
						}
					}
				},
				"description": "Get the most recent exercise data for given exercise IDs.\nReturns a map of exerciseId -> most recent exercise data from past sessions.",
				"tags": [
					"SessionLogs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"exerciseIds": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"exerciseIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/api/sessions/generate": {
			"post": {
				"operationId": "Generate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneratedSession"
								}
							}
						}
					}
				},
				"description": "Generate a personalised climbing session based on the provided inputs.",
				"tags": [
					"Sessions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SessionInput"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}